feat(propagation): limit the number of keys parsed from tracestate
# Motivation

We want to limit the amount of memory we allocate when parsing a tracestate, in order to prevent bad imputs to amplify DOS attacks by causing OOM crashes.
A user could theorically create a lot of small keys and allocate lots of small objects this way that need to be added to a hashmap

# Changes

* Move TraceState, TraceParent and InjectTraceState to the tracecontext module as this is where they belong
* Take only the first 32 entries of tracestate when parsing it.
