+++ datadog-opentelemetry/src/propagation/context.rs
@@ -7,14 +7,12 @@
 //! propagated across service boundaries, including trace IDs, span IDs,
 //! sampling decisions, and W3C tracestate data.
 
-use std::{borrow::Cow, collections::HashMap, str::FromStr, vec};
+use std::collections::HashMap;
 
-use crate::{
-    core::{
-        configuration::TracePropagationStyle,
-        sampling::{SamplingMechanism, SamplingPriority},
-    },
-    dd_debug,
+pub use super::tracecontext::{InjectTraceState, Tracestate};
+use crate::core::{
+    configuration::TracePropagationStyle,
+    sampling::{SamplingMechanism, SamplingPriority},
 };
 
 use super::tracecontext::TRACESTATE_KEY;
@@ -154,202 +152,6 @@ pub struct SpanContext {
     pub is_remote: bool,
     /// W3C tracestate information.
     pub tracestate: Option<Tracestate>,
-}
-
-/// Tracestate data to be injected into outgoing requests.
-///
-/// Contains non-Datadog tracestate entries from the parent span that should
-/// be propagated to downstream services.
-pub struct InjectTraceState {
-    header: String,
-}
-
-impl InjectTraceState {
-    pub(crate) fn from_header(header: String) -> Self {
-        Self { header }
-    }
-
-    pub(crate) fn additional_values(&self) -> impl Iterator<Item = &str> {
-        self.header.split(',').filter(|part| {
-            let (key, value) = part.split_once('=').unwrap_or((part, ""));
-            key != "dd"
-                && !value.is_empty()
-                && Tracestate::valid_key(key)
-                && Tracestate::valid_value(value)
-        })
-    }
-}
-
-#[derive(Clone, Debug, PartialEq)]
-pub(crate) struct Traceparent {
-    pub sampling_priority: SamplingPriority,
-    pub trace_id: u128,
-    pub span_id: u64,
-}
-
-/// Parsed W3C tracestate header containing Datadog-specific and vendor values.
-///
-/// The tracestate header allows vendors to propagate additional trace context
-/// alongside the standard traceparent header.
-#[derive(Clone, Default, Debug, PartialEq)]
-pub struct Tracestate {
-    pub(crate) sampling: Option<Sampling>,
-    pub(crate) origin: Option<String>,
-    pub(crate) lower_order_trace_id: Option<String>,
-    pub(crate) propagation_tags: Option<HashMap<String, String>>,
-    pub(crate) additional_values: Option<Vec<(String, String)>>,
-}
-
-/// Code inspired, and copied, by OpenTelemetry Rust project.
-/// <https://github.com/open-telemetry/opentelemetry-rust/blob/main/opentelemetry/src/trace/span_context.rs>
-impl Tracestate {
-    fn valid_key(key: &str) -> bool {
-        if key.len() > 256 {
-            return false;
-        }
-
-        let allowed_special = |b: u8| b == b'_' || b == b'-' || b == b'*' || b == b'/';
-        let mut vendor_start = None;
-        for (i, &b) in key.as_bytes().iter().enumerate() {
-            if !(b.is_ascii_lowercase() || b.is_ascii_digit() || allowed_special(b) || b == b'@') {
-                return false;
-            }
-
-            if i == 0 && (!b.is_ascii_lowercase() && !b.is_ascii_digit()) {
-                return false;
-            } else if b == b'@' {
-                if vendor_start.is_some() || i + 14 < key.len() {
-                    return false;
-                }
-                vendor_start = Some(i);
-            } else if let Some(start) = vendor_start {
-                if i == start + 1 && !(b.is_ascii_lowercase() || b.is_ascii_digit()) {
-                    return false;
-                }
-            }
-        }
-
-        true
-    }
-
-    fn valid_value(value: &str) -> bool {
-        if value.len() > 256 {
-            return false;
-        }
-
-        !(value.contains(',') || value.contains('='))
-    }
-}
-
-impl FromStr for Tracestate {
-    type Err = String;
-    fn from_str(tracestate: &str) -> Result<Self, Self::Err> {
-        let ts_v = tracestate.split(',');
-
-        let mut dd: Option<HashMap<String, String>> = None;
-        let mut additional_values = vec![];
-
-        for v in ts_v {
-            let (key, value) = v.split_once('=').unwrap_or(("", ""));
-
-            if !Tracestate::valid_key(key) || value.is_empty() || !Tracestate::valid_value(value) {
-                dd_debug!("Tracestate: invalid key or header value: {v}");
-                return Err(String::from("Invalid tracestate"));
-            }
-
-            if key == "dd" {
-                dd = Some(
-                    value
-                        .trim()
-                        .split(';')
-                        .filter_map(|item| {
-                            if !item.as_bytes().iter().all(|c| matches!(c, b' '..=b'~')) {
-                                None
-                            } else {
-                                let mut parts = item.splitn(2, ':');
-                                Some((
-                                    parts.next()?.to_string(),
-                                    decode_tag_value(parts.next()?).to_string(),
-                                ))
-                            }
-                        })
-                        .collect(),
-                );
-            } else {
-                additional_values.push((key.to_string(), value.to_string()));
-            }
-        }
-
-        let mut tracestate = Tracestate {
-            sampling: None,
-            origin: None,
-            lower_order_trace_id: None,
-            propagation_tags: None,
-            additional_values: None,
-        };
-
-        // the original order must be maintained
-        if !additional_values.is_empty() {
-            tracestate.additional_values = Some(additional_values);
-        }
-
-        let propagation_tags = if let Some(dd) = dd {
-            let mut tags = HashMap::new();
-            let mut priority = None;
-            let mut mechanism = None;
-
-            for (k, v) in dd {
-                match k.as_str() {
-                    "s" => {
-                        if let Ok(p_sp) = SamplingPriority::from_str(&v) {
-                            priority = Some(p_sp);
-                        }
-                    }
-                    "o" => tracestate.origin = Some(v),
-                    "p" => tracestate.lower_order_trace_id = Some(v.to_string()),
-                    "t.dm" => {
-                        if let Ok(p_sm) = SamplingMechanism::from_str(&v) {
-                            mechanism = Some(p_sm);
-                        }
-                        tags.insert(k, v);
-                    }
-                    _ => {
-                        tags.insert(k, v);
-                    }
-                }
-            }
-
-            tracestate.sampling = Some(Sampling {
-                priority,
-                mechanism,
-            });
-
-            Some(tags)
-        } else {
-            dd_debug!("No `dd` value found in tracestate");
-            None
-        };
-
-        tracestate.propagation_tags = propagation_tags;
-
-        Ok(tracestate)
-    }
-}
-
-fn decode_tag_value(value: &str) -> Cow<'_, str> {
-    if value.as_bytes().contains(&b'~') {
-        Cow::Owned(value.replace('~', "="))
-    } else {
-        Cow::Borrowed(value)
-    }
-}
-
-pub(crate) fn encode_tag_value(tag: &str) -> Cow<'_, str> {
-    if tag.as_bytes().contains(&b'=') {
-        Cow::Owned(tag.replace('=', "~"))
-    } else {
-        Cow::Borrowed(tag)
-    }
 }
 
 pub(crate) fn split_trace_id(trace_id: u128) -> (Option<u64>, u64) {
